l2_dotprops: make sure int64 is used (Windows issue)

+++ fafbseg/flywire/l2.py
@@ -426,7 +426,7 @@ def l2_dotprops(root_ids, min_size=None, sample=False, omit_failures=None,
     if not navis.utils.is_iterable(root_ids):
         root_ids = [root_ids]
 
-    root_ids = np.asarray(root_ids)
+    root_ids = np.asarray(root_ids, dtype=np.int64)
 
     if '0' in root_ids or 0 in root_ids:
         raise ValueError('Unable to produce dotprops for root ID 0.')
